payslip-backend/requirements.txt
authlib
fastapi
fastapi-jwt-auth
uvicorn
sqlalchemy
python-dotenv
pydantic
pydantic[email]
pydantic-settings
structlog
alembic
psycopg2-binary
asyncpg
databases[asyncpg]
httpx[http2]
python-stdnum
fastapi-jwt
mailtrap
pypdf
reportlab
reportlab[pycairo]
python-multipart
python-jose[cryptography] 
passlib[argon2]
argon2-cffi
aiofiles
uvicorn
pandas
idemptx